Separete scheduled component and tasks
